finished with refactor, all tests passing
